+++ src/lib/components/ExplorerLink.tsx
@@ -22,7 +22,8 @@ export type LinkType =
   | "code_id"
   | "block_height"
   | "proposal_id"
-  | "pool_id";
+  | "pool_id"
+  | "task_id";
 
 interface ExplorerLinkProps extends FlexProps {
   value: string;
@@ -78,6 +79,9 @@ export const getNavigationUrl = ({
       break;
     case "pool_id":
       url = "/pools";
+      break;
+    case "task_id":
+      url = "/my-module-verifications";
       break;
     case "invalid_address":
       return "";