+++ src/lib/components/module/ModuleCard.tsx
@@ -4,35 +4,32 @@ import { useMemo } from "react";
 
 import { AppLink } from "../AppLink";
 import { CustomIcon } from "../icon";
+import { MoveVerifyBadge } from "../MoveVerifyBadge";
 import { AmpEvent, track } from "lib/amplitude";
 import { ModuleInteractionMobileStep } from "lib/pages/interact/types";
-import { useMoveVerifyInfo } from "lib/services/verification/move";
-import type { BechAddr, IndexedModule, Option } from "lib/types";
+import type { MoveVerifyInfoResponse } from "lib/services/types";
+import type { IndexedModule, Option } from "lib/types";
+import { resolveMoveVerifyStatus } from "lib/utils";
 
 import { CountBadge } from "./CountBadge";
 
 interface ModuleCardProps {
-  selectedAddress: BechAddr;
   module: IndexedModule;
   selectedModule: Option<IndexedModule>;
+  moveVerifyInfo: Option<MoveVerifyInfoResponse>;
   setSelectedModule?: (module: IndexedModule) => void;
   setStep?: (step: ModuleInteractionMobileStep) => void;
   readOnly?: boolean;
 }
 
 export const ModuleCard = ({
-  selectedAddress,
   module,
+  moveVerifyInfo,
   selectedModule,
   setSelectedModule,
   setStep,
   readOnly = false,
 }: ModuleCardProps) => {
-  const { data: isVerified } = useMoveVerifyInfo(
-    selectedAddress,
-    module.moduleName
-  );
-
   const handleModuleClick = (clickedModule: IndexedModule) => {
     track(AmpEvent.USE_MODULE_CARD, {
       viewCount: clickedModule.viewFunctions.length,
@@ -65,13 +62,12 @@ export const ModuleCard = ({
           <Text className="ellipsis" variant="body2">
             {module.moduleName}
           </Text>
-          {isVerified && (
-            <CustomIcon
-              name="check-circle-solid"
-              color="success.main"
-              boxSize={3}
-            />
-          )}
+          <MoveVerifyBadge
+            status={resolveMoveVerifyStatus(
+              module.digest,
+              moveVerifyInfo?.digest
+            )}
+          />
         </Flex>
         <Flex gap={1} ml="auto">
           <CountBadge count={module.viewFunctions.length} variant="view" />
@@ -83,8 +79,8 @@ export const ModuleCard = ({
       </Grid>
     ),
     [
-      isVerified,
       JSON.stringify(module),
+      JSON.stringify(moveVerifyInfo),
       selectedModule?.moduleName,
       setSelectedModule,
     ]