+++ src/lib/layout/navbar/index.tsx
@@ -5,6 +5,7 @@ import type { Dispatch, SetStateAction } from "react";
 import {
   useCurrentChain,
   useEvmConfig,
+  useInitiaL1,
   useMoveConfig,
   usePublicProjectConfig,
   useTierConfig,
@@ -19,6 +20,7 @@ import { CollapseNavMenu } from "./Collapse";
 import { ExpandNavMenu } from "./Expand";
 import type { MenuInfo } from "./types";
 import {
+  getDeviceSubmenuMove,
   getDeviceSubmenuWasm,
   getDevSubmenuMove,
   getDevSubmenuWasm,
@@ -40,6 +42,8 @@ const Navbar = observer(({ isExpand, setIsExpand }: NavbarProps) => {
   const wasm = useWasmConfig({ shouldRedirect: false });
   const move = useMoveConfig({ shouldRedirect: false });
   const evm = useEvmConfig({ shouldRedirect: false });
+  const isInitiaL1 = useInitiaL1({ shouldRedirect: false });
+
   const { getSavedPublicProjects } = usePublicProjectStore();
   const publicProject = usePublicProjectConfig({ shouldRedirect: false });
   const isCurrentPage = useIsCurrentPage();
@@ -78,6 +82,7 @@ const Navbar = observer(({ isExpand, setIsExpand }: NavbarProps) => {
                           slug: "/saved-accounts",
                           icon: "admin" as IconKeys,
                         },
+                        ...getDeviceSubmenuMove(isInitiaL1),
                         ...getDeviceSubmenuWasm(wasm.enabled),
                       ],
                     },