+++ src/lib/pages/my-module-verification-details/types.ts
@@ -1,0 +1,5 @@
+import z from "zod";
+
+export const zMyModuleVerificationDetailsQueryParams = z.object({
+  taskId: z.string().uuid(),
+});