+++ src/lib/services/types/verification/move/index.ts
@@ -1,9 +1,9 @@
 import { z } from "zod";
 
-import { zHexAddr } from "lib/types";
-import { snakeToCamel } from "lib/utils";
+import { zHexAddr, zUtcDate } from "lib/types";
+import { mergeModulePath, snakeToCamel } from "lib/utils";
 
-enum MoveVerifyTaskStatus {
+export enum MoveVerifyTaskStatus {
   Finished = "FINISHED",
   NotFound = "NOT_FOUND",
   Pending = "PENDING",
@@ -11,30 +11,33 @@ enum MoveVerifyTaskStatus {
 }
 
 export const zSubmitMoveVerifyResponse = z.object({
-  id: z.string(),
+  id: z.string().uuid(),
   status: z.nativeEnum(MoveVerifyTaskStatus),
 });
 export type SubmitMoveVerifyResponse = z.infer<
   typeof zSubmitMoveVerifyResponse
 >;
 
+const zMoveVerificationModuleIdentifier = z.object({
+  name: z.string(),
+  address: zHexAddr,
+});
+export type MoveVerificationModuleIdentifier = z.infer<
+  typeof zMoveVerificationModuleIdentifier
+>;
+
 export const zMoveVerifyByTaskIdResponse = z.object({
   task: z.object({
-    id: z.string(),
+    id: z.string().uuid(),
     status: z.nativeEnum(MoveVerifyTaskStatus),
   }),
-  results: z
+  result: z
     .object({
-      moduleIdentifiers: z.array(
-        z.object({
-          name: z.string(),
-          address: zHexAddr,
-        })
-      ),
+      moduleIdentifiers: z.array(zMoveVerificationModuleIdentifier),
       chainId: z.string(),
       verifiedAt: z.coerce.date(),
     })
-    .optional(),
+    .nullable(),
 });
 export type MoveVerifyByTaskIdResponse = z.infer<
   typeof zMoveVerifyByTaskIdResponse
@@ -44,7 +47,7 @@ export const zMoveVerifyInfoResponse = z
   .object({
     module_address: zHexAddr,
     module_name: z.string(),
-    verified_at: z.string(),
+    verified_at: zUtcDate,
     digest: z.string(),
     source: z.string(),
     base64: z.string(),
@@ -53,13 +56,20 @@ export const zMoveVerifyInfoResponse = z
   .transform(snakeToCamel);
 export type MoveVerifyInfoResponse = z.infer<typeof zMoveVerifyInfoResponse>;
 
-export const zMoveVerifyInfosByAddressResponse = z.object({
-  contracts: z.array(
-    zMoveVerifyInfoResponse.innerType().extend({
-      id: z.string(),
-    })
-  ),
-});
+export const zMoveVerifyInfosByAddressResponse = z
+  .object({
+    contracts: z.array(zMoveVerifyInfoResponse),
+  })
+  .transform((val) =>
+    val.contracts.reduce<Record<string, MoveVerifyInfoResponse>>(
+      (acc, contract) => {
+        acc[mergeModulePath(contract.moduleAddress, contract.moduleName)] =
+          contract;
+        return acc;
+      },
+      {}
+    )
+  );
 export type MoveVerifyInfosByAddressResponse = z.infer<
   typeof zMoveVerifyInfosByAddressResponse
 >;