+++ src/lib/stores/root.ts
@@ -5,7 +5,7 @@ import { ContractStore } from "./contract";
 import { NetworkStore } from "./networks";
 import { PublicProjectStore } from "./project";
 import { SchemaStore } from "./schema";
-import { VerifyModuleTaskStore } from "./verify-module";
+import { MoveVerifyTaskStore } from "./verify-module";
 
 export class RootStore {
   accountStore: AccountStore;
@@ -22,7 +22,7 @@ export class RootStore {
 
   localChainConfigStore: LocalChainConfigStore;
 
-  verifyModuleTaskStore: VerifyModuleTaskStore;
+  moveVerifyTaskStore: MoveVerifyTaskStore;
 
   constructor() {
     this.accountStore = new AccountStore();
@@ -32,6 +32,6 @@ export class RootStore {
     this.schemaStore = new SchemaStore();
     this.networkStore = new NetworkStore();
     this.localChainConfigStore = new LocalChainConfigStore();
-    this.verifyModuleTaskStore = new VerifyModuleTaskStore();
+    this.moveVerifyTaskStore = new MoveVerifyTaskStore();
   }
 }